+++ methodical-core/src/main/java/org/jwcarman/methodical/ParameterInfo.java
@@ -13,13 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jwcarman.methodical.param;
+package org.jwcarman.methodical;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Parameter;
 import java.lang.reflect.Type;
 import java.util.Optional;
-import org.jwcarman.methodical.Named;
 import org.jwcarman.specular.TypeRef;
 
 /**