+++ methodical-core/src/test/java/org/jwcarman/methodical/DefaultMethodInvokerAccessibilityTest.java
@@ -22,7 +22,6 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import org.jwcarman.methodical.param.ParameterResolver;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class DefaultMethodInvokerAccessibilityTest {