+++ methodical-core/src/test/java/org/jwcarman/methodical/ParameterInfoTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jwcarman.methodical.param;
+package org.jwcarman.methodical;
 
 import static org.assertj.core.api.Assertions.assertThat;